Update Master with Beta

+++ scripts/build-ui.sh
@@ -2,6 +2,8 @@
 
 set -x
 
+export CI_ENVIRONMENT_NAME=production
+
 ROOT="$PWD"
 frontendFolder="$ROOT/bfx-report-ui"
 pathToTriggerElectronLoad="$frontendFolder/src/utils/triggerElectronLoad.js"
@@ -58,33 +60,42 @@ if ! [ -s "$frontendFolder/package.json" ]; then
   exit 1
 fi
 
-cp -f "$frontendFolder/.env.example" "$frontendFolder/.env"
-sed -i -e \
-  "s/REACT_APP_ELECTRON=.*/REACT_APP_ELECTRON=true/g" \
-  $frontendFolder/.env
-
-npm i --no-audit
-
 sed -i -e \
   "s/API_URL: .*,/API_URL: \'http:\/\/localhost:34343\/api\',/g" \
-  $frontendFolder/src/var/config.js
+  $frontendFolder/src/config.js
 sed -i -e \
   "s/WS_ADDRESS: .*,/WS_ADDRESS: \'ws:\/\/localhost:34343\/ws\',/g" \
-  $frontendFolder/src/var/config.js
+  $frontendFolder/src/config.js
 
 if [ $isDevEnv != 0 ]; then
+  export CI_ENVIRONMENT_NAME=development
+
 	sed -i -e \
-    "s/KEY_URL: .*,/KEY_URL: \'https:\/\/test.bitfinex.com\/api\',/g" \
-    $frontendFolder/src/var/config.js
+    "s/KEY_URL: .*,/KEY_URL: \'https:\/\/api.staging.bitfinex.com\/api\',/g" \
+    $frontendFolder/src/config.js
 fi
 
 sed -i -e \
-  "s/showAuthPage: .*,/showAuthPage: true,/g" \
-  $frontendFolder/src/var/config.js
+  "s/localExport: false/localExport: true/g" \
+  $frontendFolder/src/config.js
 sed -i -e \
-  "s/showFrameworkMode: .*,/showFrameworkMode: true,/g" \
-  $frontendFolder/src/var/config.js
+  "s/showAuthPage: false/showAuthPage: true/g" \
+  $frontendFolder/src/config.js
+sed -i -e \
+  "s/isElectronApp: false/isElectronApp: true/g" \
+  $frontendFolder/src/config.js
+sed -i -e \
+  "s/showFrameworkMode: false/showFrameworkMode: true/g" \
+  $frontendFolder/src/config.js
+
+rm -f "$ROOT/.eslintrc"
+
+npm i --no-audit
 npm run build
+
+if ! [ -s "$frontendFolder/build/index.html" ]; then
+  exit 1
+fi
 
 mv -f $frontendFolder/build/* $uiBuildFolder
 cp $pathToTriggerElectronLoad $uiBuildFolder/triggerElectronLoad.js
